# malloc: various changes and optimizations