--- a/pkg/common/malloc/c_allocator.go
+++ b/pkg/common/malloc/c_allocator.go
@@ -30,10 +30,10 @@ func NewCAllocator() *CAllocator {
 
 var _ Allocator = new(CAllocator)
 
-func (c *CAllocator) Allocate(size uint64) (unsafe.Pointer, Deallocator) {
+func (c *CAllocator) Allocate(size uint64) (unsafe.Pointer, Deallocator, error) {
 	ptr := C.malloc(C.ulong(size))
 	clear(unsafe.Slice((*byte)(ptr), size))
-	return ptr, c
+	return ptr, c, nil
 }
 
 var _ Deallocator = new(CAllocator)