--- a/pkg/fileservice/io_entry.go
+++ b/pkg/fileservice/io_entry.go
@@ -90,7 +90,10 @@ func CacheOriginalData(r io.Reader, data []byte, allocator CacheDataAllocator) (
 
 func (i *IOEntry) prepareData() (finally func(err *error)) {
 	if cap(i.Data) < int(i.Size) {
-		ptr, dec := getMallocAllocator().Allocate(uint64(i.Size))
+		ptr, dec, err := getMallocAllocator().Allocate(uint64(i.Size))
+		if err != nil {
+			panic(err)
+		}
 		metric.FSMallocLiveObjectsIOEntryData.Inc()
 		i.Data = unsafe.Slice((*byte)(ptr), i.Size)
 		if i.releaseData != nil {