## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [x] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue #16667 

## What this PR does / why we need it:
more optimizations for malloc profile

performance comparison
```
goos: linux
goarch: amd64
pkg: github.com/matrixorigin/matrixone/pkg/common/malloc
cpu: AMD Ryzen 9 7900 12-Core Processor             
                                        │      b1      │                  b2                   │
                                        │    sec/op    │    sec/op     vs base                 │
DefaultAllocator/allocate:_n_64-24        222.0n ± ∞ ¹   120.1n ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/parallel:_n_64-24        30.90n ± ∞ ¹   29.84n ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/allocate:_n_4096-24      256.1n ± ∞ ¹   139.8n ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/parallel:_n_4096-24      34.70n ± ∞ ¹   32.21n ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/allocate:_n_2097152-24   400.2n ± ∞ ¹   254.6n ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/parallel:_n_2097152-24   801.9n ± ∞ ¹   803.9n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                   164.2n         122.0n        -25.66%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                        │      b1      │                  b2                  │
                                        │     B/op     │    B/op      vs base                 │
DefaultAllocator/allocate:_n_64-24        27.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/parallel:_n_64-24        27.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/allocate:_n_4096-24      27.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/parallel:_n_4096-24      27.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/allocate:_n_2097152-24   27.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
DefaultAllocator/parallel:_n_2097152-24   28.000 ± ∞ ¹   1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                    27.16         1.000        -96.32%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                        │     b1      │                b2                │
                                        │  allocs/op  │  allocs/op   vs base             │
DefaultAllocator/allocate:_n_64-24        1.000 ± ∞ ¹   0.000 ± ∞ ¹  ~ (p=1.000 n=1) ²
DefaultAllocator/parallel:_n_64-24        1.000 ± ∞ ¹   0.000 ± ∞ ¹  ~ (p=1.000 n=1) ²
DefaultAllocator/allocate:_n_4096-24      1.000 ± ∞ ¹   0.000 ± ∞ ¹  ~ (p=1.000 n=1) ²
DefaultAllocator/parallel:_n_4096-24      1.000 ± ∞ ¹   0.000 ± ∞ ¹  ~ (p=1.000 n=1) ²
DefaultAllocator/allocate:_n_2097152-24   1.000 ± ∞ ¹   0.000 ± ∞ ¹  ~ (p=1.000 n=1) ²
DefaultAllocator/parallel:_n_2097152-24   1.000 ± ∞ ¹   0.000 ± ∞ ¹  ~ (p=1.000 n=1) ²
geomean                                   1.000                      ?               ³ ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ summaries must be >0 to compute geomean
⁴ ratios must be >0 to compute geomean

```